muda Long para Integer

+++ src/main/java/com/devsuperior/DesafioORM/entities/Atividade.java
@@ -12,7 +12,7 @@ public class Atividade {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
+    private Integer id;
 
     private String nome;
 
@@ -34,18 +34,18 @@ public class Atividade {
     public Atividade() {
     }
 
-    public Atividade(Long id, String nome, String descricao, Double preco) {
+    public Atividade(Integer id, String nome, String descricao, Double preco) {
         this.id = id;
         this.nome = nome;
         this.descricao = descricao;
         this.preco = preco;
     }
 
-    public Long getId() {
+    public Integer getId() {
         return id;
     }
 
-    public void setId(Long id) {
+    public void setId(Integer id) {
         this.id = id;
     }
