+++ src/main/java/com/devsuperior/DesafioORM/entities/Participante.java
@@ -12,7 +12,7 @@ public class Participante {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
+    private Integer id;
 
     private String nome;
 
@@ -30,17 +30,17 @@ public class Participante {
     private Set<Atividade> atividades = new HashSet<>();
 
 
-    public Participante(Long id, String nome, String email) {
+    public Participante(Integer id, String nome, String email) {
         this.id = id;
         this.nome = nome;
         this.email = email;
     }
 
-    public Long getId() {
+    public Integer getId() {
         return id;
     }
 
-    public void setId(Long id) {
+    public void setId(Integer id) {
         this.id = id;
     }
